Move core option manager into retroarch.c
