+++ griffin/griffin.c
@@ -141,7 +141,6 @@ CONFIG FILE
 
 #include "../libretro-common/file/config_file.c"
 #include "../libretro-common/file/config_file_userdata.c"
-#include "../managers/core_option_manager.c"
 
 /*============================================================
 RUNTIME FILE
