+++ managers/core_option_manager.h
@@ -48,71 +48,13 @@ struct core_option_manager
 typedef struct core_option_manager core_option_manager_t;
 
 /**
- * core_option_manager_new:
- * @conf_path        : Filesystem path to write core option config file to.
- * @vars             : Pointer to variable array handle.
+ * core_option_manager_set_default:
+ * @opt                   : pointer to core option manager object.
+ * @idx                   : index of core option to be reset to defaults.
  *
- * Creates and initializes a core manager handle.
- *
- * Returns: handle to new core manager handle, otherwise NULL.
+ * Reset core option specified by @idx and sets default value for option.
  **/
-core_option_manager_t *core_option_manager_new(const char *conf_path,
-      const void *data);
-
-/**
- * core_option_manager_updated:
- * @opt              : options manager handle
- *
- * Has a core option been updated?
- *
- * Returns: true (1) if a core option has been updated,
- * otherwise false (0).
- **/
-bool core_option_manager_updated(core_option_manager_t *opt);
-
-/**
- * core_option_manager_flush:
- * @opt              : options manager handle
- *
- * Writes core option key-pair values to file.
- *
- * Returns: true (1) if core option values could be
- * successfully saved to disk, otherwise false (0).
- **/
-bool core_option_manager_flush(core_option_manager_t *opt);
-
-/**
- * core_option_manager_flush_game_specific:
- * @opt              : options manager handle
- * @path             : path for the core options file
- *
- * Writes core option key-pair values to a custom file.
- *
- * Returns: true (1) if core option values could be
- * successfully saved to disk, otherwise false (0).
- **/
-bool core_option_manager_flush_game_specific(
-      core_option_manager_t *opt, const char* path);
-
-/**
- * core_option_manager_free:
- * @opt              : options manager handle
- *
- * Frees core option manager handle.
- **/
-void core_option_manager_free(core_option_manager_t *opt);
-
-void core_option_manager_get(core_option_manager_t *opt,  void *data);
-
-/**
- * core_option_manager_size:
- * @opt              : options manager handle
- *
- * Gets total number of options.
- *
- * Returns: Total number of options.
- **/
-size_t core_option_manager_size(core_option_manager_t *opt);
+void core_option_manager_set_default(core_option_manager_t *opt, size_t idx);
 
 /**
  * core_option_manager_get_desc:
@@ -140,36 +82,6 @@ const char *core_option_manager_get_val(core_option_manager_t *opt,
 
 void core_option_manager_set_val(core_option_manager_t *opt,
       size_t idx, size_t val_idx);
-
-/**
- * core_option_manager_next:
- * @opt                   : pointer to core option manager object.
- * @idx                   : idx of core option to be reset to defaults.
- *
- * Get next value for core option specified by @idx.
- * Options wrap around.
- **/
-void core_option_manager_next(core_option_manager_t *opt, size_t idx);
-
-/**
- * core_option_manager_prev:
- * @opt                   : pointer to core option manager object.
- * @idx                   : idx of core option to be reset to defaults.
- * Options wrap around.
- *
- * Get previous value for core option specified by @idx.
- * Options wrap around.
- **/
-void core_option_manager_prev(core_option_manager_t *opt, size_t idx);
-
-/**
- * core_option_manager_set_default:
- * @opt                   : pointer to core option manager object.
- * @idx                   : idx of core option to be reset to defaults.
- *
- * Reset core option specified by @idx and sets default value for option.
- **/
-void core_option_manager_set_default(core_option_manager_t *opt, size_t idx);
 
 RETRO_END_DECLS
